requirements.txt
# py -3 -m venv venv
# [pip Requirements Files](https://pip.pypa.io/en/latest/user_guide/#requirements-files)

# venv/Scripts/activate.bat
# Select interpreter. Kill / new Bash terminal.
# whidh pip3
# py -m pip install -r requirements.txt --upgrade

pip
setuptools
wheel

# [filetype](https://pypi.org/project/filetype/)
filetype=>1.0.13

# [Geocoding API](https://developers.google.com/maps/documentation/geocoding)
googlemaps=>4.6.0

# [pylibtiff](https://github.com/pearu/pylibtiff)
libtiff=>0.4.2

# [NumPy](https://numpy.org/)
numpy=>1.23.1

# [pathlib](https://docs.python.org/3/library/pathlib.html)
  
# [Pillow](https://pillow.readthedocs.io/en/stable/)
# [Pillow: IFDRational is not JSON serializable #6199](https://github.com/python-pillow/Pillow/issues/6199)
Pillow=>9.2.0

# [python-decouple](https://github.com/henriquebastos/python-decouple/)
python-decouple=>3.6

# [requests](https://pypi.org/project/requests/)
requests=>2.28.1

# [rich](https://rich.readthedocs.io/en/stable/introduction.html)
rich=>12.4.4

# [typer](https://typer.tiangolo.com/)
typer=>0.5.0

# [urllib3](https://urllib3.readthedocs.io/en/stable/)
urllib3=>1.26.10

